Refactor beta and gamma tower bursts into modules
